special token tokenization handle
