+++ Makefile
@@ -6,7 +6,7 @@ LDFLAGS	=	-lreadline -lncurses
 
 
 SRCS	=	main.c \
-			./tokenizer/tokenizer_utils.c ./tokenizer/tokenizer.c ./tokenizer/token_maker.c ./tokenizer/bool_funcs.c \
+			./tokenizer/tokenizer_utils.c ./tokenizer/tokenizer.c ./tokenizer/token_maker.c \
 			./expander/expander.c ./expander/expand_helpers.c ./expander/expand_handler.c \
 			./signals/signals.c \
 			./redirection/redirs.c redirection/redir_free.c ./redirection/redir_apply.c ./redirection/redir_helpers.c \
@@ -14,7 +14,7 @@ SRCS	=	main.c \
 			./syntax/syntax_utils.c ./syntax/syntax.c \
 			./ast/ast.c ./ast/ast_helper.c ./ast/ast_redir.c ./ast/ast_tools.c ./ast/ast_cmd.c\
 			./env/env_parser.c ./env/env_utils.c ./env/parsing_helper.c \
-			./execute/execute.c ./execute/cmd_exec.c ./execute/builtin_exec.c ./execute/utils.c ./execute/exec_utils.c ./execute/pipe_exec.c\
+			./execute/execute.c ./execute/cmd_exec.c ./execute/builtin_exec.c ./execute/utils.c ./execute/exec_utils.c ./execute/pipe_exec.c \
 			./built_in/cd.c ./built_in/echo.c ./built_in/env.c ./built_in/export.c\
 			./built_in/export_utils.c ./built_in/exit.c\
 			./built_in/pwd.c ./built_in/unset.c ./built_in/utils.c\
