--- tokenizer/bool_funcs.c
@@ -1,34 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   bool_funcs.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: apatvaka <apatvaka@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/07/16 20:10:31 by rbarkhud          #+#    #+#             */
-/*   Updated: 2025/08/16 23:14:39 by apatvaka         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "tokenizer.h"
-
-int	ft_isspace(char c)
-{
-	return ((c >= 9 && c <= 13) || c == ' ');
-}
-
-int	ft_inset(char c, char *set)
-{
-	int	i;
-
-	if (!set)
-		return (0);
-	i = 0;
-	while (set[i])
-	{
-		if (c == set[i])
-			return (1);
-		++i;
-	}
-	return (0);
-}
