+++ tokenizer/token_maker.c
@@ -6,7 +6,7 @@
 /*   By: rbarkhud <rbarkhud@student.42yerevan.am    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/17 16:52:28 by rbarkhud          #+#    #+#             */
-/*   Updated: 2025/09/27 01:46:34 by rbarkhud         ###   ########.fr       */
+/*   Updated: 2025/11/05 22:13:26 by rbarkhud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ static int	quote_handler(char *str, int i)
 
 int	make_word_token(t_token **head, char *str, int i)
 {
-	int		j;
+	int	j;
 
 	j = 0;
 	while (str[i + j])
@@ -50,12 +50,14 @@ int	make_word_token(t_token **head, char *str, int i)
 
 int	make_specials_token(t_token **head, char *str, int i)
 {
-	int	j;
+	int		j;
+	char	prev;
 
 	j = 0;
+	prev = str[i];
 	while (str[i + j])
 	{
-		if (ft_inset(str[i + j], "|&<>"))
+		if (str[i + j] == prev && ft_inset(str[i + j], "|&<>") && j < 2)
 			++j;
 		else
 			break ;
@@ -63,4 +65,32 @@ int	make_specials_token(t_token **head, char *str, int i)
 	if (j != 0)
 		add_token(head, ft_substr(str, i, j));
 	return (i + j);
+}
+
+void	add_token(t_token **head, char *value)
+{
+	t_token	*temp;
+	t_token	*new_node;
+
+	if (!head || !value || !(*value))
+		return ;
+	new_node = (t_token *)malloc(sizeof(t_token));
+	if (!new_node)
+	{
+		free(value);
+		return ;
+	}
+	new_node->next = NULL;
+	new_node->token_type = get_token_type(value);
+	new_node->token = value;
+	new_node->is_subshell = false;
+	if (!(*head))
+		(*head) = new_node;
+	else
+	{
+		temp = (*head);
+		while (temp->next)
+			temp = temp->next;
+		temp->next = new_node;
+	}
 }
